+++ src/main/java/com/boot/model/cluster/JsEbookTemplateExample.java
@@ -404,7 +404,7 @@ public class JsEbookTemplateExample {
             return (Criteria) this;
         }
 
-        public Criteria andRtypeEqualTo(Byte value) {
+        public Criteria andRtypeEqualTo(Integer value) {
             addCriterion("rtype =", value, "rtype");
             return (Criteria) this;
         }
